Showing keyboard for input in list editor when FAB is clicked
